+++ react/CurrentPrice.tsx
@@ -1,0 +1,3 @@
+import CurrentPrice from './components/CurrentPrice'
+
+export default CurrentPrice