+++ react/services/useProductWithBenefits.ts
@@ -1,6 +1,8 @@
 import { useProduct } from 'vtex.product-context'
 import type { ProductContextState } from 'vtex.product-context/react/ProductTypes'
 
+import { getDefaultSeller } from '../helpers/seller'
+
 interface Benefit {
   teaserType?: string
   items?: Array<{
@@ -21,17 +23,22 @@ type ProductContextWithBenefits =
 
 export const useProductWithBenefits = () => {
   const productContextValue = useProduct() as ProductContextWithBenefits
-
   const benefits = productContextValue?.product?.benefits?.filter(
     (b) => b?.teaserType === 'Catalog'
   )
 
+  const productName = productContextValue?.product?.productName
   const selectedItem = productContextValue?.selectedItem
+  const seller = getDefaultSeller(selectedItem?.sellers)
+  const commertialOffer = seller?.commertialOffer
+
   const measurementUnit = selectedItem?.measurementUnit ?? ''
-  const price = selectedItem?.sellers?.[0]?.commertialOffer?.Price
-  const teasers = selectedItem?.sellers?.[0]?.commertialOffer?.teasers
+  const price = commertialOffer?.Price
+  const teasers = commertialOffer?.teasers
 
   return {
+    productName,
+    selectedItem,
     benefits,
     measurementUnit,
     price,